Fix linux error

+++ gamebase/Player.hpp
@@ -38,7 +38,7 @@ public:
 	int GetLevel() {return cur_level_;}
     int GetCurLevelExp() {return exp_;}
     int GetCurLevelMaxExp() {
-		if (GetLevel() < level_exps_.size()) {
+		if (GetLevel() < (int)level_exps_.size()) {
 			return level_exps_[GetLevel()];
 		}
 		else {
